# Optimize sorting for non-cray networks