The current sorting code uses `unorderedCopy` for small messages. This
idiom is only optimized for Cray networks (comm=ugni) and is very slow
on other platforms.

This PR adds initial support for aggregating copies (`dst=src`) and uses
it for sorting on non-ugni configurations. There is a lot more that can
be done to optimize aggregation, and longer-term aggregation will
outperform unorderedCopy under ugni as well. For now, I just wanted to
get some initial aggregation committed since it has such a large benefit
for non-cray networks.

Here are some sort performance results for an FDR InfiniBand network
with gn-ibv-large in MB/s per node. This is the benchmark from #162 with
`./sort-perf --positiveValues=true --maxBits=16 --memFraction=50`

| nodes | unordered   | aggregated |
| ----- | ----------- | ---------- |
|   1   | 520.00 MB/s | 565 MB/s   |
|   2   |   1.65 MB/s | 420 MB/s   |
|   4   |   0.36 MB/s | 350 MB/s   |
|   8   |   0.34 MB/s | 320 MB/s   |
|  16   |   0.49 MB/s | 300 MB/s   |
|  32   |   0.53 MB/s | 250 MB/s   |

Performance is also significantly improved for Ethernet networks. I
don't have a real Ethernet cluster to run on, but I was able to use the
out-of-band Ethernet on that same IB cluster. At 16 nodes, performance
goes from 0.01 MB/s to 8 MB/s. This is not a production network and I
would expect results to be much better on a real Ethernet cluster.

There is a non-trivial memory overhead for aggregation. Each task has
per-destination buffers and uses `bufferSize * numLocales * elemType`
space. Sorting `int`s on 32 nodes with 36 cores per node this requires
~128 MB per node for aggregation buffers. At 512 nodes this would
require ~2 GB per node.