@ronawho 2019-11-11
This will probably be easiest to review ignoring whitespace -- https://github.com/mhmerrill/arkouda/pull/185/files?w=1

This is a cleaned-up version of the prototype from https://github.com/mhmerrill/arkouda/issues/162#issuecomment-549540542. I dropped the default buffer size from 32,768 to 4,096. This hurts performance but lessons the memory overhead a lot.

Over time I plan to tune more and limit memory overheads, but I think this is a good start. If you want to experiment on your IB clusters, the important knob to tune is `dstBuffSize`.